convert to byte and short to int so it works on SQLServer

+++ src/Umbraco.Cms.Persistence.EFCore.SqlServer/Migrations/UmbracoDbContextModelSnapshot.cs
@@ -496,8 +496,8 @@ namespace Umbraco.Cms.Persistence.EFCore.SqlServer.Migrations
                         .HasColumnType("nvarchar(max)")
                         .HasColumnName("thumbnail");
 
-                    b.Property<byte>("Variations")
-                        .HasColumnType("tinyint")
+                    b.Property<int>("Variations")
+                        .HasColumnType("int")
                         .HasColumnName("variations");
 
                     b.HasKey("PrimaryKey");
@@ -836,8 +836,8 @@ namespace Umbraco.Cms.Persistence.EFCore.SqlServer.Migrations
                         .HasColumnType("datetime2")
                         .HasColumnName("createDate");
 
-                    b.Property<short>("Level")
-                        .HasColumnType("smallint")
+                    b.Property<int>("Level")
+                        .HasColumnType("int")
                         .HasColumnName("level");
 
                     b.Property<Guid?>("NodeObjectType")
@@ -973,8 +973,8 @@ namespace Umbraco.Cms.Persistence.EFCore.SqlServer.Migrations
                         .HasColumnType("nvarchar(500)")
                         .HasColumnName("validationRegExpMessage");
 
-                    b.Property<byte>("Variations")
-                        .HasColumnType("tinyint")
+                    b.Property<int>("Variations")
+                        .HasColumnType("int")
                         .HasColumnName("variations");
 
                     b.HasKey("Id");
@@ -1019,10 +1019,10 @@ namespace Umbraco.Cms.Persistence.EFCore.SqlServer.Migrations
                         .HasColumnType("nvarchar(max)")
                         .HasColumnName("text");
 
-                    b.Property<short>("Type")
+                    b.Property<int>("Type")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("smallint")
-                        .HasDefaultValue((short)0)
+                        .HasColumnType("int")
+                        .HasDefaultValue(0)
                         .HasColumnName("type");
 
                     b.Property<Guid>("UniqueId")
@@ -1185,10 +1185,10 @@ namespace Umbraco.Cms.Persistence.EFCore.SqlServer.Migrations
                         .HasColumnType("uniqueidentifier")
                         .HasColumnName("key");
 
-                    b.Property<short>("Kind")
+                    b.Property<int>("Kind")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("smallint")
-                        .HasDefaultValue((short)0)
+                        .HasColumnType("int")
+                        .HasDefaultValue(0)
                         .HasColumnName("kind");
 
                     b.Property<DateTime?>("LastLockoutDate")
